Separate request-processing from routes into controllers
