+++ server/src/controllers/404.controller.ts
@@ -1,10 +1,10 @@
 import { Request, Response } from "express";
 
-const notFoundRouter = (req: Request, res: Response) => {
+const notFound = (req: Request, res: Response) => {
     res.status(404).json({
         error: 404,
         message: "Route not found."
     });
 };
 
-export default notFoundRouter;
+export default notFound;
