+++ server/src/routes/posts.ts
@@ -2,68 +2,17 @@
 /* Catch clause variable type annotation must be 'any' or 'unknown' if specified.ts(1196) */
 
 import express from "express";
-import {
-    createPost,
-    getPostsByMonastic,
-    deletePost
-} from "../models/posts.model";
-import { Post } from "../_db/schemas";
-import logger from "../_utils/logger";
+import { createPostPost, monasticPostGet, monasticPostDelete } from "../controllers/posts.controller";
 
 const router = express.Router();
 
-// Create a new post
-router.post("/", async (req, res) => {
-    try {
-        const post: Post = req.body;
-        /*  TODO handle logic for uploading to S3 
-            and creating a media entry with the URI */
-        await createPost(post);
+// POST create a new post
+router.post("/", createPostPost);
 
-        res.status(201).json({
-            success: true,
-            message: "Post created successfully.",
-            post
-        });
-    } catch (error: any) {
-        logger.error("Error creating post:", error.message);
-        res.status(error.status ?? 500).json({
-            success: false,
-            error: error.message
-        });
-    }
-});
+// GET posts for a specific monastic
+router.get("/:monasticId", monasticPostGet);
 
-// Get posts for a specific monastic
-router.get("/:monasticId", async (req, res) => {
-    try {
-        const { monasticId } = req.params;
-        const posts = await getPostsByMonastic(monasticId);
-
-        res.json({ success: true, posts });
-    } catch (error: any) {
-        logger.error("Error getting posts:", error.message);
-        res.status(error.status ?? 500).json({
-            success: false,
-            error: error.message
-        });
-    }
-});
-
-// Delete a post
-router.delete("/:postId", async (req, res) => {
-    try {
-        const { postId } = req.params;
-        await deletePost(postId);
-
-        res.json({ success: true, message: "Post deleted successfully." });
-    } catch (error: any) {
-        logger.error("Error deleting post:", error.message);
-        res.status(error.status ?? 500).json({
-            success: false,
-            error: error.message
-        });
-    }
-});
+// DELETE a post
+router.delete("/:postId", monasticPostDelete);
 
 export default router;
